requirements.txt
tqdm
python-dateutil
glob2
requests
